Cleaned up almost all the compiler warnings.
git-svn-id: https://svn.aps.anl.gov/AOP/oag/trunk/apps/src/elegant@21775 c1b61890-85f3-ea29-e4d6-a272fe79f168

+++ makeInitialParticleEnsemble.c
@@ -100,7 +100,7 @@ int makeInitialParticleEnsemble(
     memcpy((*final)[i], (*initial)[i], COORDINATES_PER_PARTICLE*sizeof(double));
   }
   
-  printf("%ld duplicates were removed from initial ensemble, leaving %ld vectors\n", n_duplicates, n_points_total);
+  printf("%ld duplicates were removed from initial ensemble, leaving %ld vectors\n", (long)n_duplicates, (long)n_points_total);
   
   return n_points_total;
 }
